// fronts/admin/src/app/interfaces/administrative.ts
export interface SchoolInterface {
    id: number;
    country_id: number;
    city_id: number;
    school_id: string;
    school_name: string;
    dni: string;
    director_name: string;
    suburb: string;
    address: string;
    location: string;
    phones: string;
    web: string;
    email: string;
    number_locations: number;
}